requirements.txt
# Core
numpy>=1.24
anndata>=0.10
scanpy>=1.10

# Classical VAE models
scvi-tools>=1.2

# Foundation models (optional — install only what you need)
# pip install scgpt
# pip install geneformer
# pip install uce

# Shared by scvi-tools and all foundation models
torch>=2.0
